+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src/View.js
@@ -1,6 +1,0 @@
-
-Ext.define('Fiddle.View', {
-    extend: 'Ext.panel.Panel',
-    html: 'Ext version ~ ' + Ext.getVersion().version
-});
-