+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src/meta.js
@@ -1,6 +1,6 @@
 var meta = {
-    fiddleHeader: 'fiddle-0036-ItunesTop100',
-    fiddleSubHeader: 'Template Fiddle created @ 09-29-16' +
+    fiddleHeader: 'Ext Itunes',
+    fiddleSubHeader: 'Open Source Itunes <i>like</i> UI' +
     '<br />'
 };