+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src/view/Viewport.js
@@ -1,0 +1,8 @@
+
+Ext.define('tunes.view.Viewport', {
+    extend: 'tunes.view.MyViewport',
+    requires: [
+        'tunes.view.MyViewport',
+        'tunes.view.PreviewWindow'
+    ]
+});